Remove dist

+++ kicad_amf_plugin/utils/platebtn.py
@@ -192,7 +192,7 @@ class PlateButton(wx.Control):
             bw, bh = bmp.GetSize()
             ypos = (self.GetSize()[1] - bh) // 2
             if tw > 0:
-                gc.DrawBitmap(bmp, (width - tw)/2 -  bw -6 , ypos, bmp.GetMask() is not None)
+                gc.DrawBitmap(bmp, int((width - tw)/2 -  bw -6) , ypos, bmp.GetMask() is not None)
             else:
                 gc.DrawBitmap(bmp, 6 , ypos, bmp.GetMask() is not None)
 
@@ -367,9 +367,9 @@ class PlateButton(wx.Control):
             # else:
             #     gc.DrawText(self.Label, txt_x + 2, txt_y)
             if wx.Platform == '__WXGTK__':
-                dc.DrawText(self.Label, (width - tw) /2 ,txt_y)
+                dc.DrawText(self.Label, int((width - tw) /2) ,txt_y)
             else:
-                gc.DrawText(self.Label, (width - tw) /2, txt_y)
+                gc.DrawText(self.Label, int((width - tw) /2), txt_y)
             self.__DrawDropArrow(gc, width - 10, (height // 2) - 2)
